+++ meta/recipes-update/mender-state-scripts-base/files/300_Inhibit-reboot-script-standalone.sh
@@ -1,0 +1,6 @@
+#!/bin/sh
+
+PREFIX=INHIBIT-REBOOT-SCRIPT-STANDALONE
+
+echo "Inhibiting reboot script for standalone mode"
+touch /data/mender/upgrade/inhibit-reboot-script-standalone