+++ meta/recipes-update/mender-state-scripts-base/files/999_Final-cleanup.sh
@@ -1,5 +1,8 @@
 #!/bin/sh
 
+source /data/mender/upgrade/migration-env.sh
+
 PREFIX=FINAL-CLEANUP
 
-rm -rf /data/mender/upgrade
+log $PREFIX "Cleaning up"
+rm -rf $MENDER_UPGRADE_DIR